"Refactor navbar-logo class to small-logo and Add footer
Renamed the class 'navbar-logo' to 'small-logo' in the home.css file to improve the clarity and purpose of the class name. Corresponding references in index.html also updated for consistency.

A new footer has also been added in the index.html. This provides a navigation advantage for the users and includes additional information such as links to different sections, a logo image, copyright message, and a disclaimer."