"Implement interaction improvements for user navigation
Implemented an interactive scroll button that quickly navigates users to the top of the page when clicked. To further enhance user experience, a disappearing navigation bar has been added which automatically hides while scrolling down and reappears when scrolling up.

These changes aim to boost usability by providing cleaner, more intuitive navigation. The scroll button is hidden when the user is at the top of the page, ensuring interface elements do not obstruct the view when not necessary. The disappearing navigation bar contributes towards a more immersive, distraction-free browsing experience.

Additionally, it was necessary to adjust corresponding CSS and Javascript files to support the new features. Renamed class 'navbar-logo' to 'small-logo' in home.css for more accurate representation and also modulated content in index.html to maintain consistency across changes."